Moved tool functions and conversion function invocation to ToolsManager and FunctionRegistry as part of mdenet/educationplatform#40
