+++ platform/src/ToolsManager.js
@@ -1,21 +1,25 @@
-import { parseConfigFile } from "./Utility.js";
+import { parseConfigFile, ARRAY_ANY_ELEMENT } from "./Utility.js";
 import { FunctionRegistry } from "../src/FunctionRegistry.js"
 import { ActionFunction } from "./ActionFunction.js";
 import { ToolConfigValidator } from "./ToolConfigValidator.js";
 import { EducationPlatformError } from "./EducationPlatformError.js";
 
 class ToolManager {
-
     toolId;
     toolsUrls;
     configErrors = [];
     configValidator;
     tools = {};
     functionRegister;
+    errorNotification;
 
-    constructor(){
+    /**
+     * @param {function(String)} errorNotifier - the function to call to display an error
+     */
+    constructor(errorNotifier){
         this.configValidator= new ToolConfigValidator();
-        this.functionRegister= new FunctionRegistry();
+        this.functionRegister= new FunctionRegistry(this);
+        this.errorNotification = errorNotifier;
     }
 
     setToolsUrls(urls){
@@ -40,7 +44,6 @@ class ToolManager {
     }
 
     
-
     /** 
      * Fetches the tools from the tools url and populates tool's
      * functions and panel definitions.
@@ -220,8 +223,6 @@ class ToolManager {
     }
 
 
-
-
     /**
      * Finds the action function for an action function Id
      * @param {*} actionFuntionId 
@@ -241,18 +242,6 @@ class ToolManager {
         console.log("Tool with function id '" + actionFuntionId + "' was not found.");
         return null;
 
-    }
-
-
-    /**
-     * Resolves the id of an action function to the function itself
-     * 
-     * TODO: To be moved to the FunctionRegistry issue #40
-     * 
-     * @param {string} functionId 
-     */
-    functionRegistry_resolve(functionId){
-        return this.getActionFunction(functionId)
     }
 
     
@@ -310,28 +299,249 @@ class ToolManager {
     }
 
 
-    /**
-     * Returns the id of the registered function with matching input and output parameters.
-     * @param {*} inputsParamTypes 
-     * @param {*} outputParamType 
-     * @returns The id of a matching function
-     */
-    getConversionFunction(inputsParamTypes, outputParamType){
-        
-       return this.functionRegister.lookupFunction(inputsParamTypes, outputParamType);
-    }
-
-    getPartiallyMatchingConversionFunctions(inputsParamTypes, outputParamType){
-        
-        return this.functionRegister.lookupFunctionsPartialMatch(inputsParamTypes, outputParamType);
-     }
-
    /**
      * Returns the errors found parsing and validating the tool configuration files 
      * @returns array of errors
      */
     getConfigErrors(){
         return this.configErrors;
+    }
+
+
+    /**
+     * Invokes an action function by placing requests to all the required external tool functions  
+     * 
+     * @param {string} functionId the id of tool function
+     * @param {Map} parameterMap map from parameter name to its value and type
+     * @returns {Promise}  promise to result of the action function 
+     */
+    invokeActionFunction(functionId, parameterMap){
+
+        let actionFunction = this.functionRegister.resolve(functionId);
+
+        let parameterPromises = [];
+
+        for ( const paramName  of parameterMap.keys() ){ /* TODO add defensive checks that every required value
+                                                                is provided issue #57 */
+
+            let actionFunctionParam = actionFunction.getParameters().find( p => p.name === paramName);                                                             
+                                                                        
+            /* Check the given parameter types against the those of the requested action function. 
+            If required, request conversion from available tool functions */
+            let givenParameter = parameterMap.get(paramName);
+
+            if (givenParameter.type != actionFunctionParam.type){
+                //Types don't match  so try  and convert 
+                let convertedValue;
+
+                const metamodelId = actionFunction.getInstanceOfParamName(paramName);
+                
+                if(metamodelId==null){
+                    // Convert with no metamodel to consider
+                    convertedValue = this.convert( givenParameter.value, givenParameter.type, 
+                                            actionFunctionParam.type, paramName ); // TODO issue #58 remove paramName
+
+                } else {
+                    // Convert considering metamodel
+                    const givenMetamodel = parameterMap.get(metamodelId);
+
+                    convertedValue = this.convertIncludingMetamodel( givenParameter.value , givenParameter.type, 
+                                                                givenMetamodel.value, givenMetamodel.type, 
+                                                                actionFunctionParam.type, paramName ); // TODO issue #58 remove paramName
+                }
+
+                parameterPromises.push(convertedValue);
+            
+            } else {
+                // Matching types add values to promise for synchronisation 
+                let value =  new Promise( function (resolve) { 
+                    let parameterData = {};
+                    
+                    parameterData.name = paramName;
+                    parameterData.data = givenParameter.value;
+        
+                    resolve(parameterData); 
+                });
+
+                parameterPromises.push(value);
+            }
+        }
+
+        // Invoke the actionFunction on compeletion of any conversions
+        let actionFunctionPromise = new Promise((resolve, reject) => {
+
+            Promise.all( parameterPromises ).then( (values) => { 
+                let actionRequestData = {};
+
+                //Populate the transformed parameters
+                for ( const param  of actionFunction.getParameters() ){
+
+                    const panelConfig = parameterMap.get(param.name); 
+
+                    if (panelConfig == undefined){
+                        // Set unused parameters in the request to undefined as the epsilon backend function expects them all. 
+                        actionRequestData[param.name] = "undefined";
+
+                    } else {
+                        let parameterData = values.find(val => (val.name === param.name) );
+
+                        actionRequestData[param.name] =  parameterData.data;
+                    }
+                }
+
+                let resultPromise = this.functionRegister.call(functionId, actionRequestData);
+
+                resolve(resultPromise);
+            
+            }).catch( (err) => {
+
+                reject(err);
+            });
+
+        });
+
+        return actionFunctionPromise;
+    }
+
+    /**
+     * Converts a source value to a target type using the available conversion functions
+     * 
+     * @param {string} sourceValue 
+     * @param {string} sourceType 
+     * @param {string} targetType
+     * @param {string} parameterName name of the parameter for request
+     * @returns {Promise} promise for the converted paramter value
+     */
+    convert(sourceValue, sourceType, targetType, parameterName){
+        
+        let parameterPromise;
+        let typesPanelValuesMap = {}; // Types have to be distinct for mapping to the conversion function's paramters
+        typesPanelValuesMap[sourceType]=  sourceValue;
+
+        let conversionFunctionId = this.functionRegister.find( Object.keys(typesPanelValuesMap), targetType );
+
+        if (conversionFunctionId != null){
+            //There is a matching conversion function
+            parameterPromise = this.functionRegister.callConversion(conversionFunctionId, typesPanelValuesMap, parameterName);
+            
+        } else {
+            parameterPromise = null;
+
+            this.errorNotification("No conversion function available for input types:" + Object.keys(typesPanelValuesMap).toString() )
+        }
+
+        return parameterPromise;
+    }
+
+
+    /**
+     * Converts a source value to a target type using the available conversion functions taking
+     * into consideration the related metamodel.
+     * 
+     *   TODO: To be moved to the ToolManager issue #40
+     * 
+     * @param {string} sourceValue 
+     * @param {string} sourceType
+     * @param {string} metamodelValue 
+     * @param {string} metamodelType
+     * @param {string} targetType
+     * @param {string} parameterName name of the parameter for request
+     * @returns {Promise} promise for the converted paramter value
+     */
+    async convertIncludingMetamodel(sourceValue, sourceType, metamodelValue, metamodelType, targetType, parameterName){
+        let parameterPromise;
+        let typesPanelValuesMap = {}; // Types have to be distinct for mapping to the conversion function's paramters
+        typesPanelValuesMap[sourceType]=  sourceValue;
+
+        let conversionFunctionId;
+
+        let potentialConversionFunctions = this.functionRegister.findPartial( [sourceType, ARRAY_ANY_ELEMENT], targetType);
+
+        //check for a conversion function with the metamodel type
+        conversionFunctionId = await this.selectConversionFunctionConvertMetamodel( metamodelType, metamodelValue, potentialConversionFunctions, 
+                                                                                false, parameterName, typesPanelValuesMap)
+
+        if (conversionFunctionId==null){
+            //no conversion found so check for a conversion function but consider conversions of the metamodel
+            conversionFunctionId = await this.selectConversionFunctionConvertMetamodel(metamodelType, metamodelValue, potentialConversionFunctions, 
+                                                                                    true, parameterName, typesPanelValuesMap);
+        }
+
+        if (conversionFunctionId != null){
+            //There is a matching conversion function
+            parameterPromise = this.functionRegister.callConversion(conversionFunctionId, typesPanelValuesMap, parameterName);
+            
+        } else {
+            parameterPromise = null;
+            
+            this.errorNotification("No conversion function available for input types:" + Object.keys(typesPanelValuesMap).toString() )
+        }
+
+        return parameterPromise;
+    }
+
+
+    /**
+     * For the given list of conversion function ids to check, finds the first conversion function with matching metamodel dependency.
+     * Optionally conversions of the metamodel are considered from the conversion functions available to the tools manager and
+     * the metamodel type. If available, the metamodel value is converted to the required type. 
+     * 
+     * @param {string} metamodelType the metamodel type
+     * @param {string} metamodelValue the metamodel value
+     * @param {string[]} conversionFunctions list of conversion function ids to check 
+     * @param {boolean} convertMetamodel when true try to convert the metamodel using a remote tool service conversion function
+     *                                    available to the ToolsManager.
+     * @param {string} parameterName the name of the parameter to use when convering the metamodel. 
+     * @param {string[]} typeValueMap the type values map the metamodel input value is added to if a conversion function is found
+     * @returns {string} the id of a conversion function to use, null if none found.
+     */
+    async selectConversionFunctionConvertMetamodel(metamodelType, metamodelValue, conversionFunctions, convertMetamodel, parameterName, typeValueMap){
+        let conversionFunctionId;
+        let functionsToCheck = [...conversionFunctions]
+        
+        while ( conversionFunctionId==null && functionsToCheck.length > 0){
+            let functionId = functionsToCheck.pop();
+            let conversionFunction = this.getActionFunction(functionId);
+
+            // Lookup the conversion function's metamodel type
+            let metamodelName = conversionFunction.getInstanceOfParamName( conversionFunction.getParameters()[0].name );
+
+            if(metamodelName==null){
+                metamodelName = conversionFunction.getInstanceOfParamName( conversionFunction.getParameters()[1].name );
+            }
+
+            const targetMetamodelType = conversionFunction.getParameterType(metamodelName);
+
+            if (!convertMetamodel){
+                // Check for conversion functions with matching metamodels only
+                
+                if (targetMetamodelType==metamodelType) {
+                        //Conversion function found so use the panel value
+                        
+                        conversionFunctionId = functionId;
+                        typeValueMap[metamodelType]=  metamodelValue;
+                }
+
+            } else {
+                // Check for conversion functions converting metamodel if possible 
+                let metamodelConversionFunctionId = this.functionRegister.find( [metamodelType], targetMetamodelType );
+                
+                if (metamodelConversionFunctionId != null){
+
+                    conversionFunctionId = functionId;
+
+                    //convert metamodel
+                    let metamodelTypeValueMap =  {};  
+                    metamodelTypeValueMap[metamodelType]=metamodelValue; // The found conversion function is expected to have one parameter
+
+                    let convertedValue = await this.functionRegister.callConversion(metamodelConversionFunctionId, metamodelTypeValueMap, parameterName);
+
+                    typeValueMap[targetMetamodelType]= convertedValue.data;
+                }
+            }
+        }
+
+        return conversionFunctionId;
     }
 }
